+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v0.2.0 (in development)
 - Support for weekend events added
 - Day-of-week abbreviations in input files may no longer be lowercase
 - Added a `--no-weekends` option for not showing Sunday & Saturday
+- Added a `--start-monday` option for starting the week on Monday
 
 v0.1.0 (2019-04-11)
 -------------------